chore(builtin): adding more tests for slice builtin
