+++ tests/unittests/resources/DiagnosticsSuite/runtime/slice_of_step_0.ark
@@ -1,0 +1,1 @@
+(slice "abc" 0 2 0)
