+++ tests/unittests/resources/DiagnosticsSuite/runtime/slice_out_of_range.ark
@@ -1,0 +1,1 @@
+(slice "abc" 4 5)
