+++ tests/unittests/resources/DiagnosticsSuite/typeChecking/slice_num_num_num.ark
@@ -1,0 +1,1 @@
+(slice 1 5 5)
