+++ tests/unittests/resources/DiagnosticsSuite/typeChecking/slice_str_str_num.ark
@@ -1,0 +1,1 @@
+(slice "abc" "def" 5)
