+++ tests/unittests/resources/LangSuite/builtins-tests.ark
@@ -43,4 +43,18 @@
         (test:eq (slice "abcdef" nil nil 2) "ace")
         (test:eq (slice "abcdef" nil nil -2) "fdb")
         (test:eq (slice "abcdef" 1 nil 2) "bdf")
-        (test:eq (slice "abcdef" 1 nil -2) "fdb") }) })
+        (test:eq (slice "abcdef" 1 nil -2) "fdb")
+
+        (test:eq (slice [1 2 3] nil nil) [1 2 3])
+        (test:eq (slice [1 2 3] nil 1) [1])
+        (test:eq (slice [1 2 3] nil 2) [1 2])
+        (test:eq (slice [1 2 3] nil 3) [1 2 3])
+        (test:eq (slice [1 2 3] nil 30) [1 2 3])
+        (test:eq (slice [1 2 3 4 5 6] 1 -1) [2 3 4 5])
+        (test:eq (slice [1 2 3 4 5 6] 1 -1 30) [2])
+        (test:eq (slice [1 2 3 4 5 6] 1 -1 -1) [5 4 3 2])
+        (test:eq (slice [1 2 3 4 5 6] 1 -1 -30) [5])
+        (test:eq (slice [1 2 3 4 5 6] nil nil 2) [1 3 5])
+        (test:eq (slice [1 2 3 4 5 6] nil nil -2) [6 4 2])
+        (test:eq (slice [1 2 3 4 5 6] 1 nil 2) [2 4 6])
+        (test:eq (slice [1 2 3 4 5 6] 1 nil -2) [6 4 2]) }) })
